- removed some '#if 0'ed code - changed a few instances where a variable was being set to the same value twice in a row for no reason - removed an instance(s) where a variable was set to itself - removed some places where an if/else block was doing the exact same thing in either condition - removed OurBool and changed it to a proper bool type - removed RGBALIGHT_MAKE as it was just a macro that directed straight to RGBA_MAKE (which I used to replace all instances of RGBA_LIGHT with)
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@435

+++ 3dc/OnScreenKeyboard.cpp
@@ -202,7 +202,7 @@ void Osk_Draw()
 	int pos_x = osk_x + indent_space;
 	int pos_y = osk_y + indent_space;
 
-	uint32_t index = 0;
+	size_t index = 0;
 
 	for (uint32_t y = 0; y < numVerticalKeys; y++)
 	{
